Make possible multiple C-R-like files
Change the function that creates the path for challenge-response
storate to take a "suffix" argument, so we can have multiple files.
Isolate the part of the challenge-response forming function that
fills in the expected response into a callback, so we now can store
other things in the "expected response" field.