Enable saved encrypted authtoken (password) p.2
Implement function in the PAM module, when given the argument
"supply_authtoken", to extract and decode data stored by
ykpamcfg action "add_saved_password" and supply it as
PAM_AUTHTOK to subsequent PAM modules. Particularly to
pam_gnome_keyring module (and equivalents) to decode the
keyring if yubikey was used as the sole authentication factor.